feat: utilities for instantiating Babylon contracts (#66)
This PR provides a message `MsgInstantiateBabylonContracts` for
instantiating all Babylon contracts. This simplifies the Cosmos
integration: one can first store code of 3 contracts, and then submit
this message for instantiating everything.

With this message as baseline, we can then change the handler logic to
improve the instantiation flow, e.g., instantiate 3 contracts one by one
in this handler, rather than allowing contracts instantiating each
other. This will be done in a subsequent PR.

The PR contains the following:

- [x] function/message for instantiating Babylon contracts
- [x] CLIs for the messages
- [x] fuzz tests
- [x] e2e tests