refactor: improve contracts storage & remove init
The current module implementation implements a method
for receiving contract code IDs and performing the instantiation of them.
This method is limited as it introduced a hard dependency of the
module to the instantiation methods of the contract,
while making the module more complicated.

This PR modifies the code base to:
* Have a message `SetBSNContracts that instead of instantiating the contracts,
  receives their address and sets them in internal storage.
* Create a new type that tightly couples the contracts together
  as they need to be set and retrieved at the same time.
* Remove the contracts from the parameters in order to have only a single
  place in which governance can set the contracts
  (instead of both the `SetBSNContracts` message and params modification).
  This leads to modifications on the genesis state and a new storage entry.
* Modify the tests and demo app and all peripheral codes to deal with this.