refactor: improve contracts storage & remove init (#165)
The current module implementation implements a method for receiving
contract code IDs and performing the instantiation of them. This method
is limited as it introduced a hard dependency of the module to the
instantiation methods of the contract, while making the module more
complicated.

This PR modifies the code base to:
* Have a message `SetBSNContracts that instead of instantiating the
contracts, receives their address and sets them in internal storage.
* Create a new type that tightly couples the contracts together as they
need to be set and retrieved at the same time.
* Remove the contracts from the parameters in order to have only a
single place in which governance can set the contracts (instead of both
the `SetBSNContracts` message and params modification). This leads to
modifications on the genesis state and a new storage entry.
* Modify the tests and demo app and all peripheral codes to deal with
this.

Resolves #136 and #114 

The new flow now for people using the SDK is:
1. Upload the four contract codes
2. Instantiate the Babylon contract, which in turn will instantiate the
other 3 contracts
3. Collect the addresses of the four contracts and submit a
`MsgSetBSNContracts` message coming from `Authority` (in most cases
governance, i.e., a governance proposal for gov to execute the
`MsgSetBSNContracts` is needed).

---------

Co-authored-by: Fangyu Gai <greferry@gmail.com>